Refactor 'list'; Move 'listForUser' to 'libs'
