+++ services/notes/libs/list-lib.js
@@ -1,0 +1,21 @@
+const listForUser = async (userId) => {
+  const params = {
+    TableName: process.env.tableName,
+    // 'KeyConditionExpression' defines the condition for the query
+    // - 'userId = :userId': only return items with matching 'userId'
+    //   partition key
+    KeyConditionExpression: "userId = :userId",
+    // 'ExpressionAttributeValues' defines the value in the condition
+    // - ':userId': defines 'userId' to be the id of the author
+    ExpressionAttributeValues: {
+      ":userId": userId,
+    },
+  };
+
+  const result = await dynamoDb.query(params);
+
+  // Return the matching list of items in response body
+  return result.Items;
+};
+
+export { listForUser };
